requirements.txt
fastapi
requests
python-dotenv
uvicorn
python-dotenv
beautifulsoup4
markdown
requests<3
duckdb
numpy
python-dateutil
docstring-parser
httpx
scikit-learn
pydantic
requests
flask
pandas
whisper
openai
faker
markdown
duckdb
bs4
Pillow